Remove old template paths
