+++ Server/templates/base.html
@@ -15,11 +15,11 @@
             <a href="{{ url_for('scan.hello') }}" class="text-xl font-semibold">WareEye</a>
             <div class="hidden md:block ml-10 space-x-4">
               <a href="{{ url_for('scan.hello') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'scan.hello' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Home</a>
-              <a href="{{ url_for('scan.list_scans') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'scan.list_scans' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Scans</a>
-              <a href="{{ url_for('scan.list_destination_codes') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'scan.list_destination_codes' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Destination Codes</a>
-              <a href="{{ url_for('scan.list_dock_doors') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'scan.list_dock_doors' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Dock Doors</a>
-              <a href="{{ url_for('scan.list_olpn_labels') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'scan.list_olpn_labels' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">OLPN Labels</a>
-              <a href="#" class="px-3 py-2 rounded-md text-sm font-medium text-gray-300 hover:bg-primary/80 hover:text-white">Stream / Live Feed</a>
+              <a href="{{ url_for('facility.list_scans') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'facility.list_scans' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Scans</a>
+              <a href="{{ url_for('facility.list_destination_codes') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'facility.list_destination_codes' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Destination Codes</a>
+              <a href="{{ url_for('facility.list_dock_doors') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'facility.list_dock_doors' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Dock Doors</a>
+              <a href="{{ url_for('facility.list_olpn_labels') }}" class="px-3 py-2 rounded-md text-sm font-medium {% if request.endpoint == 'facility.list_olpn_labels' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">OLPN Labels</a>
+              <a href="{{ url_for('operations.loading_dashboard') }}" class="px-3 py-2 rounded-md text-sm font-medium text-gray-300 hover:bg-primary/80 hover:text-white">Stream / Live Feed</a>
             </div>
           </div>
           <div class="md:hidden">
@@ -33,11 +33,11 @@
       </div>
       <div id="mobile-menu" class="hidden md:hidden px-2 pt-2 pb-3 space-y-1">
         <a href="{{ url_for('scan.hello') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'scan.hello' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Home</a>
-        <a href="{{ url_for('scan.list_scans') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'scan.list_scans' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Scans</a>
-        <a href="{{ url_for('scan.list_destination_codes') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'scan.list_destination_codes' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Destination Codes</a>
-        <a href="{{ url_for('scan.list_dock_doors') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'scan.list_dock_doors' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Dock Doors</a>
-        <a href="{{ url_for('scan.list_olpn_labels') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'scan.list_olpn_labels' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">OLPN Labels</a>
-        <a href="#" class="block px-3 py-2 rounded-md text-base font-medium text-gray-300 hover:bg-primary/80 hover:text-white">Stream / Live Feed</a>
+        <a href="{{ url_for('facility.list_scans') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'facility.list_scans' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Scans</a>
+        <a href="{{ url_for('facility.list_destination_codes') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'facility.list_destination_codes' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Destination Codes</a>
+        <a href="{{ url_for('facility.list_dock_doors') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'facility.list_dock_doors' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">Dock Doors</a>
+        <a href="{{ url_for('facility.list_olpn_labels') }}" class="block px-3 py-2 rounded-md text-base font-medium {% if request.endpoint == 'facility.list_olpn_labels' %}border-b-2 border-primary text-white{% else %}text-gray-300 hover:bg-primary/80 hover:text-white{% endif %}">OLPN Labels</a>
+        <a href="{{ url_for('operations.loading_dashboard') }}" class="block px-3 py-2 rounded-md text-base font-medium text-gray-300 hover:bg-primary/80 hover:text-white">Stream / Live Feed</a>
       </div>
     </nav>
     <div class="container mt-4">
