+++ Server/templates/facility/destination_code_edit.html
@@ -11,6 +11,6 @@
     <input type="text" name="name" class="form-control" value="{{ code.name }}" required>
   </div>
   <button type="submit" class="btn btn-primary">Save Changes</button>
-  <a href="{{ url_for('scan.list_destination_codes') }}" class="btn btn-secondary">Cancel</a>
+  <a href="{{ url_for('facility.list_destination_codes') }}" class="btn btn-secondary">Cancel</a>
 </form>
 {% endblock %}
