+++ Server/templates/facility/dock_door_edit.html
@@ -23,6 +23,6 @@
     <label class="form-check-label" for="is_active">Active</label>
   </div>
   <button type="submit" class="btn btn-primary">Save Changes</button>
-  <a href="{{ url_for('scan.list_dock_doors') }}" class="btn btn-secondary">Cancel</a>
+  <a href="{{ url_for('facility.list_dock_doors') }}" class="btn btn-secondary">Cancel</a>
 </form>
 {% endblock %}
