+++ Server/templates/facility/olpn_label_edit.html
@@ -15,6 +15,6 @@
     </select>
   </div>
   <button type="submit" class="btn btn-primary">Save Changes</button>
-  <a href="{{ url_for('scan.list_olpn_labels') }}" class="btn btn-secondary">Cancel</a>
+  <a href="{{ url_for('facility.list_olpn_labels') }}" class="btn btn-secondary">Cancel</a>
 </form>
 {% endblock %}
