+++ Server/templates/facility/scan_edit.html
@@ -31,6 +31,6 @@
     <input type="text" name="timestamp" class="form-control" value="{{ scan.timestamp.isoformat() }}" required>
   </div>
   <button type="submit" class="btn btn-primary">Save Changes</button>
-  <a href="{{ url_for('scan.list_scans') }}" class="btn btn-secondary">Cancel</a>
+  <a href="{{ url_for('facility.list_scans') }}" class="btn btn-secondary">Cancel</a>
 </form>
 {% endblock %}
