+++ Server/templates/operations/loading_dashboard.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+{% block content %}
+<h1>Loading Dashboard</h1>
+<p>Placeholder for truck loading status.</p>
+{% endblock %}
